requirements.txt
fastapi
uvicorn
psycopg[binary]
requests
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
pypdf
python-multipart